+++ scripts/add_electricity.py
@@ -508,7 +508,7 @@ def attach_conventional_generators(
                 n.generators.loc[idx, attr] = values
 
 
-def attach_hydro(n, costs, ppl):
+def attach_hydro(n, costs, ppl, hydro_min_inflow_pu=1):
     if "hydro" not in snakemake.params.renewable:
         return
     c = snakemake.params.renewable["hydro"]
@@ -523,20 +523,12 @@ def attach_hydro(n, costs, ppl):
         .rename(index=lambda s: str(s) + " hydro")
     )
 
-    # Current fix, NaN technologies set to ROR
-    if ppl.technology.isna().any():
-        n_nans = ppl.technology.isna().sum()
-        logger.warning(
-            f"Identified {n_nans} hydro powerplants with unknown technology.\n"
-            "Initialized to 'Run-Of-River'"
-        )
-        ppl.loc[ppl.technology.isna(), "technology"] = "Run-Of-River"
-
     ror = ppl.query('technology == "Run-Of-River"')
     phs = ppl.query('technology == "Pumped Storage"')
     hydro = ppl.query('technology == "Reservoir"')
+    tbd = ppl[ppl.technology.isna()]  # To be determined technologies
 
-    inflow_idx = ror.index.union(hydro.index)
+    inflow_idx = ror.index.union(hydro.index).union(tbd.index)
     if not inflow_idx.empty:
         with xr.open_dataarray(snakemake.input.profile_hydro) as inflow:
             found_plants = ppl.ppl_id[ppl.ppl_id.isin(inflow.indexes["plant"])]
@@ -545,12 +537,16 @@ def attach_hydro(n, costs, ppl):
             # if missing time series are found, notify the user and exclude missing hydro plants
             if not missing_plants_idxs.empty:
                 # original total p_nom
-                total_p_nom = ror.p_nom.sum() + hydro.p_nom.sum()
+                total_p_nom = ror.p_nom.sum() + hydro.p_nom.sum() + tbd.p_nom.sum()
 
                 ror = ror.loc[ror.index.intersection(found_plants.index)]
                 hydro = hydro.loc[hydro.index.intersection(found_plants.index)]
+                tbd = tbd.loc[tbd.index.intersection(found_plants.index)]
+
                 # loss of p_nom
-                loss_p_nom = ror.p_nom.sum() + hydro.p_nom.sum() - total_p_nom
+                loss_p_nom = (
+                    ror.p_nom.sum() + hydro.p_nom.sum() + tbd.p_nom.sum() - total_p_nom
+                )
 
                 logger.warning(
                     f"'{snakemake.input.profile_hydro}' is missing inflow time-series for at least one bus: {', '.join(missing_plants_idxs)}."
@@ -566,6 +562,25 @@ def attach_hydro(n, costs, ppl):
                     .transpose("time", "name")
                     .to_pandas()
                 )
+
+    # Heuristics for missing hydro technologies
+    if not tbd.empty:
+        inflow_pu_limit = inflow_t[tbd.index].mean() / tbd["p_nom"]  # Average MWh/MW
+        mask_reservoir = inflow_pu_limit >= hydro_min_inflow_pu
+        mask_ror = ~mask_reservoir
+        to_be_hydro = tbd[mask_reservoir]
+        to_be_ror = tbd[mask_ror]
+
+        logger.info(
+            f"Identified {len(tbd)} hydro powerplants with unknown technology.\n"
+            f"Hydropower plants with energy-to-capacity ratio ≥ {hydro_min_inflow_pu} "
+            f"are classified as 'Reservoir'. The rest are 'Run-Of-River'.\n"
+            f"Reservoir: {mask_reservoir.sum()} \n"
+            f"Run-Of-River: {mask_ror.sum()}"
+        )
+
+        ror = pd.concat([ror, to_be_ror])
+        hydro = pd.concat([hydro, to_be_hydro])
 
     if "ror" in carriers and not ror.empty:
         n.madd(
@@ -633,13 +648,12 @@ def attach_hydro(n, costs, ppl):
         missing_countries = pd.Index(hydro["country"].unique()).difference(
             max_hours_country.dropna().index
         )
+        hydro_max_hours_default = c.get("hydro_max_hours_default", 6.0)
         if not missing_countries.empty:
             logger.warning(
-                "Assuming max_hours=6 for hydro reservoirs in the countries: {}".format(
-                    ", ".join(missing_countries)
-                )
+                f"Assuming max_hours={hydro_max_hours_default} for hydro reservoirs in the countries: "
+                "{}".format(", ".join(missing_countries))
             )
-        hydro_max_hours_default = c.get("hydro_max_hours_default", 6.0)
         hydro_max_hours = hydro.max_hours.where(
             hydro.max_hours > 0, hydro.country.map(max_hours_country)
         ).fillna(hydro_max_hours_default)
@@ -941,7 +955,9 @@ if __name__ == "__main__":
         extendable_carriers,
         snakemake.params.length_factor,
     )
-    attach_hydro(n, costs, ppl)
+    attach_hydro(
+        n, costs, ppl, snakemake.params.renewable["hydro"]["hydro_min_inflow_pu"]
+    )
     attach_existing_batteries(n, costs, ppl)
 
     if snakemake.params.electricity.get("estimate_renewable_capacities"):